Moved course specific functionality into module
